replace "daisywolrd" with current matrix version
